requirements.txt
flask
python-telegram-bot==13.7
diffusers
torch
transformers
accelerate